pyproject.toml
[tool.poetry]
name = "hybrid-search-demo"
version = "0.1.0"
description = ""
authors = ["u-masao <4973920+u-masao@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
torch = "^2.0.1"
python = "^3.12"
datasets = "^3.1.0"
click = "^8.1.7"
pandas = "^2.2.3"
transformers = "^4.46.2"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
tqdm = "^4.67.0"
mlflow = "^2.17.2"
openai = "^1.54.4"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"
dvc = "^3.56.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
